% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilitiesParsing.R
\name{simulateMeeting}
\alias{simulateMeeting}
\title{produces a fake meeting in the form of a data frame}
\usage{
simulateMeeting(seed = NULL, nSpeakers = 4, nTurns = 50, nTags = 5)
}
\arguments{
\item{seed}{random seed. Defaults to NULL}

\item{nSpeakers}{number of speakers (default: 4)}

\item{nTurns}{number of turns (default: 50)}

\item{nTags}{number of tags (default: 5)}
}
\value{
a data frame with 3 fields:
speaker
speech (random letters and punctuation)
tag
}
\description{
produces a fake meeting in the form of a data frame
}

